UnitMap thread-safety: add acquire/release code for recursive init
Replace double checked locking with C++11 acquire/release.
Cannot use CallOnce, since init is recursive and it's hard to get rid of that.
Depending on your taste, the resulting code is reasonable or ugly.

A documented gap is that init is not exception-safe (it already wasn't).
This is hard to get right, because of the 2 stage drama for the recursion.

Fall back to always lock on pre-C++11.

The cleaning up was also using double-checked locking.
Decided to better have fast lazy init and documented missing
The way the code is written means that it's hard to get this thread-safe, fast and clean without API/ABI changes.

Replace double checked locking for FITS unit extension init with a regular lock.
Looks like this code is not executed often.
If proven wrong, we can also roll some call_once or acquire/release into it.