MSTableInit thread-safety: change always lock to acquire/release scheme
Improve upon my earlier patch a few months ago to make parallel MS creation thread-safe.
Like the prev patch for UnitMap, this is the other case where init is recursive and it's difficult to get rid of that. Hence CallOnce cannot be used.

So roll our own acquire/release thing for C++11, again with the lack of exception-safety.
Half of the changed lines is reindentation.

A few APERTIF parallel MS creation test runs did not show any crashes or deadlocks. FWIW...