Make 2 caches in Quanta available for USE_THREADS=OFF and pre-C++11.
Hopefully, defining thead_local as __thread in aipsdef.h has no really nasty conseq... Otherwise, we'll have to define AIPS_TLS as either.

One issue I could not solve is the dynamic init of __thread array[50] in Euler.cc:
GCC wants a dynamic initializer for __thread (C++03), but I cannot please GCC (so exclude it).
For MVPosition, it's slightly different and not a problem.

__APPLE__ already has these tricks disabled (code was by NRAO).
I've checked various webpages, but it's vague why it's not supported. Clang does, but not the Xcode variation. Also depends on std lib.
Document it, but continue to stay away if __APPLE__. (I have no such system, so cannot easily test build it.)