Change CMake defaults: enable CXX11, USE_THREADS, USE_OPENMP, FFTW
With the new thread sync code, C++11 is highly preferable. So is USE_THREADS=ON.

If USE_THREADS, but OpenMP is N/A (may happen w/ Clang), warn but accept.
Other way around, USE_OPENMP ON, but USE_THREADS OFF is not allowed.

These 4 option defaults are already set for CASA builds.